backend/requirements.txt
aerich==0.9.1
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
asyncclick==8.3.0.7
certifi==2025.10.5
click==8.3.0
dictdiffer==0.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iso8601==2.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pypika-tortoise==0.6.2
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
sentry-sdk==2.42.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.48.0
tortoise-orm==0.25.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
asyncpg
passlib
python-jose[cryptography]
python-dotenv